Moving all config over to config file
