+++ evaluation-tasks/util/luigi.py
@@ -4,11 +4,7 @@ Common Luigi classes and functions for evaluation tasks
 
 import os
 import luigi
-
-
-def ensure_dir(dirname):
-    if not os.path.exists(dirname):
-        os.makedirs(dirname)
+import requests
 
 
 class WorkTask(luigi.Task):
@@ -46,3 +42,29 @@ class WorkTask(luigi.Task):
         d = "_workdir/%s/" % self.name
         ensure_dir(d)
         return d
+
+
+def download_file(url, local_filename):
+    """
+    The downside of this approach versus `wget -c` is that this
+    code does not resume.
+    The benefit is that we are sure if the download completely
+    successfuly, otherwise we should have an exception.
+    From: https://stackoverflow.com/a/16696317/82733
+    TODO: Would be nice to have a TQDM progress bar here.
+    """
+    # NOTE the stream=True parameter below
+    with requests.get(url, stream=True) as r:
+        r.raise_for_status()
+        with open(local_filename, "wb") as f:
+            for chunk in r.iter_content(chunk_size=8192):
+                # If you have chunk encoded response uncomment if
+                # and set chunk_size parameter to None.
+                # if chunk:
+                f.write(chunk)
+    return local_filename
+
+
+def ensure_dir(dirname):
+    if not os.path.exists(dirname):
+        os.makedirs(dirname)
